requirements.txt
opencv-python
click
pyautogui
pillow
